==> 1-Data-integration-class-balancing.ipynb <==
# Dataset :- CIA Records Search Tool Metadata - https://data.world/cia-crest-files/cia-crest-archive-metadata
# Data file :-  crest_lite_4.csv
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

# this file does following:
#1.	Merge two datasets
#2.	Remove features which are not useful
#3.	Remove null values
#4.	do sampling for class balance
#5.	merge titles into one feature and drop rest of the titles from data set
# Save the file for further processing


# CREST crest_lite_1.csv
# Classes 
#General CIA Records            96702
#NGA Records (Formerly NIMA)    35273
#Consolidated Translations      29665
#Scientific Abstracts            4347
#Library of Congress             3255
#Ground Photo Caption Cards      1094
#df = pd.read_csv('https://query.data.world/s/5iz7cqrdwstmkmrr2amlih5twvsb5m')

#CREST crest_lite_2.csv       !!!! not useful because contains only one class  -General CIA Records - 137873
#df = pd.read_csv('https://query.data.world/s/ayi3kpxolvhrf5odzzwp7mru5oaap3')

#CREST crest_lite_3.csv            !!!! 404 not found error
# df = pd.read_csv('https://query.data.world/s/u23ow7ihgemqfwate5pkuwora73asz')

#CREST crest_lite_4.csv
#df = pd.read_csv('https://query.data.world/s/4lmzkr5rrah6icpkfw4nzfigynuvsh')
#General CIA Records           54343
#Scientific Abstracts          27878
#STARGATE                       8683	
#Ground Photo Caption Cards     7879
#JPRS                           2547
#OSS Collection                  329
#Secret Writing                    2




# CREST crest_lite_1.csv
df1 = pd.read_csv('https://query.data.world/s/5iz7cqrdwstmkmrr2amlih5twvsb5m')

#CREST crest_lite_4.csv
df2 = pd.read_csv('https://query.data.world/s/4lmzkr5rrah6icpkfw4nzfigynuvsh')

df1.shape

df2.shape

# stack the DataFrames on top of each other/ Merge datasets row -wise 
df = pd.concat([df1, df2], axis=0)
df.shape

df.describe()

df.head(15)

# Keeping relevant features for further exploration

df = df[[ 'title', 'collection', 'document_number', 'release_decision', 'document_page_count',
       'sequence_number', 'publication_date' , 'content_type', 'more1_title', 'more2_title', 'more3_title',
        'more4_title', 'more5_title']]
df.shape

# Check the null values
df.isnull().sum().tolist()

# Drop null values
df = df.dropna(how="any")

# Check the null values
df.isnull().sum().tolist()

df.shape

# check for the class imbalance
df['collection'].value_counts()

# merge columns with less values than 10,000 and name it Misc
df['collection'] = df['collection'].map({'General CIA Records': 'General_CIA_Records' ,
                                       'NGA Records (Formerly NIMA)': 'NGA_Records' ,
                                       'Scientific Abstracts' : 'Scientific_Abstracts',
                                       'Consolidated Translations': 'Consolidated_Translations',
                                       'Ground Photo Caption Cards': 'Misc',
                                       'STARGATE' : 'Misc',
                                       'Library of Congress': 'Misc',
                                       'OSS Collection' : 'Misc',
                                       'JPRS' : 'Misc'})


# check for the class imbalance after merging
df['collection'].value_counts()

df.head(4)

df.tail(4)

# This is a class imbalance problem, let's do undersampling to keep all the values equal for every class

    # use resample method from scikit-learn
    from sklearn.utils import resample

    # Separate majority and minority classes for undersampling
    df_collection_General_CIA_Records              = df[df.collection== 'General_CIA_Records']
df_collection_NGA_Records                      = df[df.collection == 'NGA_Records']
df_collection_Scientific_Abstracts             = df[df.collection== 'Scientific_Abstracts']
df_collection_Consolidated_Translations        = df[df.collection== 'Consolidated_Translations']
df_collection_Misc                             = df[df.collection== 'Misc']

#df_majority = df[df.balance==0]
#df_minority = df[df.balance==1]
# Ref: https://elitedatascience.com/imbalanced-classes
 
    
df_collection_General_CIA_Records = resample(df_collection_General_CIA_Records, 
                                             replace=True,    # sample with replacement
                                             n_samples=15000,     # to match number of values in each class
                                             random_state=123) # reproducible results

df_collection_NGA_Records   = resample(df_collection_NGA_Records, 
                                             replace=True,    # sample with replacement
                                             n_samples=15000,     # to match number of values in each class
                                             random_state=123) # reproducible results

df_collection_Scientific_Abstracts = resample(df_collection_Scientific_Abstracts, 
                                             replace=True,    # sample with replacement
                                             n_samples=15000,     # to match number of values in each class
                                             random_state=123) # reproducible results

df_collection_Consolidated_Translations = resample(df_collection_Consolidated_Translations, 
                                             replace=True,    # sample with replacement
                                             n_samples=15000,     # to match number of values in each class
                                             random_state=123) # reproducible results

df_collection_Misc  = resample(df_collection_Misc, 
                                replace=True,    # sample with replacement
                                n_samples=15000,     # to match number of values in each class
                                random_state=123) # reproducible results

    
# Combine all the class with equal number of values
df_sampled = pd.concat([df_collection_General_CIA_Records, df_collection_NGA_Records, df_collection_Scientific_Abstracts,
                       df_collection_Consolidated_Translations, df_collection_Misc])
 

# check for the class balance
df_sampled['collection'].value_counts()

# change df with sampled one
df= df_sampled

df.columns

# it's better to merge all the different titles so that further can be made easy 
df[['title' , 'more1_title' , 'more2_title' , 'more3_title' , 'more4_title', 'more5_title']].head(5)

# let's merge all the title columns into title
df['title'] = df[['title' , 'more1_title' , 'more2_title' , 'more3_title' , 'more4_title', 'more5_title']].apply( 
    lambda x : ' '.join(x), axis = 1 )


pd.set_option('display.max_colwidth', -1)  # or 199
print(df['title'].head(5))

# we can drop rest of the titles
df = df[['title', 'collection', 'document_number', 'release_decision', 'document_page_count', 'sequence_number',
         'publication_date', 'content_type' ]]
df.shape

df.head(5)

# let's save the data file at this point
df.to_csv(r'datasets\crest-after-step0.csv', index=False)
