==> 2-data-separate-text-numerical.ipynb <==
# Dataset :- CIA Records Search Tool Metadata - https://data.world/cia-crest-files/cia-crest-archive-metadata
# Data file :-  crest_lite_4.csv
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt


# this file is separating data into two files - textual and numerical/categorical after converting the class labels into 
# numerical features

df = pd.read_csv(r'datasets\crest-after-step0.csv')

df.head(4)

# change collection into the numerical values with factorize() method
#df['category_id'] = df['Product'].factorize()[0]
df['collection_labels'] = df['collection'].factorize()[0]

df['collection_labels'].value_counts()

collection_labels_df = df[['collection', 'collection_labels']].drop_duplicates().sort_values('collection_labels')
collection_labels_df

# save/pickle the the collection labels dataframe for future use 
import pickle

pickle_save = open(r"pickle-intermediate-data\collection_labels_df.pickle","wb")
pickle.dump(collection_labels_df, pickle_save)
pickle_save.close()


# converting to dict 
label_dict = collection_labels_df.to_dict() 

print(label_dict)

df.shape

df.columns

df.tail(7)

# let's save the text data with labels into one file and numeric in another
data_text = df[['title' , 'collection', 'collection_labels']]

data_numeric = df[['collection', 'document_number', 'release_decision', 'document_page_count', 'sequence_number',
                   'publication_date', 'content_type', 'collection_labels']]
# Save text data file with the labels
data_text.to_csv(r'datasets\crest-data-text.csv', index=False)
# save numeric data file with the labels
data_numeric.to_csv(r'datasets\crest-data-numeric.csv', index=False)

# explore the data
# document page count
print("Number of pages in the document varies from "+ str(df['document_page_count'].max()) +
      " to "+str(df['document_page_count'].min()))


df['collection'][df['content_type'] == 'MEMO'].value_counts()

df['content_type'].value_counts()
