==> 4B-text-data-predictive-modelling.ipynb <==
# in this file we are goona work on the text data, extract features from text and build models
# Now our data cleaning has almost been done. It's time to extract more features :
# 1. n-grams
# 2. tf-idf
# 3. bag of words


import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

# let's do some moedelling
from sklearn.model_selection import train_test_split, cross_val_score

# packages for metric for evalaution of the models
from sklearn import metrics
from sklearn.metrics import classification_report , accuracy_score, confusion_matrix


from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer

# As tf–idf is very often used for text features, there is also another class called TfidfVectorizer that combines 
# all the options of CountVectorizer and TfidfTransformer in a single model.
# so we can calculate tf-idf and the CountVectorizer in one go with below module
from sklearn.feature_extraction.text import TfidfVectorizer

# use multiple models
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.svm import LinearSVC, SVC

import seaborn as sns



df = pd.read_csv(r'datasets\1a-crest-after-text-cleaning.csv')

df.head(5)

# calculating tf-idf using scikitlearn
# sublinear_df  - is set to True to use a logarithmic form for frequency.
# min_df - is the minimum numbers of documents a word must be present in to be kept.
# norm - is set to l2, to ensure all our feature vectors have a euclidian norm of 1.
# ngram_range -  is set to (1, 2) to indicate that we want to consider both unigrams and bigrams.
# stop_words - is set to "english" to remove all common pronouns ("a", "the", ...) to reduce the number of noisy features.


tfidf = TfidfVectorizer(sublinear_tf=True, min_df=50, norm='l2', encoding='latin-1', ngram_range=(1, 3))

features = tfidf.fit_transform(df.title.values.astype('str')).toarray()
labels = df.collection_labels

features.shape

features

# convert features ndarray into pandas dataframe so that it can be merged with 
#other features like avg_word_count and avg_word_len
features_df = pd.DataFrame(features)

features_df.tail(7)

df_word = df[['title_word_count' , 'avg_word_len']]

X_train, X_test, y_train, y_test = train_test_split(df_word, labels, random_state = 0)

# let's try how ensemling ( Random forest wit 10 decision trees ) performs 
random_forest = RandomForestClassifier(n_estimators=10)
random_forest.fit(X_train, y_train)

print("Accuracy = ")
print(random_forest.score(X_test, y_test))
print("\n")
y_pred_rf = random_forest.predict(X_test)
print(classification_report(y_test, y_pred_rf, target_names=df['collection'].unique()))

df_word.shape

type(df_word)

df_word.head(5)

final_features = pd.concat([features_df, df_word], axis=1)

final_features.shape

final_features.head(5)

X_train, X_test, y_train, y_test = train_test_split(final_features, labels, random_state = 0)

# Multinomial Naive Bayes
#clf = MultinomialNB().fit(X_train_tfidf, y_train)
#y_pred = clf.predict(X_test)
#print(classification_report(y_test, y_pred))


models = [
    RandomForestClassifier(n_estimators=200, max_depth=3, random_state=0),
    LinearSVC(),
    MultinomialNB(),
    LogisticRegression(random_state=0),
]

# define cross validation
CV = 5
cv_df = pd.DataFrame(index=range(CV * len(models)))
entries = []

for model in models:
  model_name = model.__class__.__name__
  accuracies = cross_val_score(model, features, labels, scoring='accuracy', cv=CV)
    
  for fold_idx, accuracy in enumerate(accuracies):
    entries.append((model_name, fold_idx, accuracy))

cv_df = pd.DataFrame(entries, columns=['model_name', 'fold_idx', 'accuracy'])
cv_df.groupby('model_name').accuracy.mean()


cv_df

# since LinearSVM is working pretty goood
# let's try different kernals and see which one is the best
models = [
    #LinearSVC(),
    SVC(kernel='poly' , degree= 2),
    SVC(kernel='rbf'),
    SVC(kernel='sigmoid')
]

# define cross validation
CV = 5
cv_df = pd.DataFrame(index=range(CV * len(models)))
entries = []

for model in models:
  model_name = model.__class__.__name__
  accuracies = cross_val_score(model, features, labels, scoring='accuracy', cv=CV)
    
  for fold_idx, accuracy in enumerate(accuracies):
    entries.append((model_name, fold_idx, accuracy))

cv_df = pd.DataFrame(entries, columns=['model_name', 'fold_idx', 'accuracy'])
cv_df.groupby('model_name').accuracy.mean()


cv_df

# it can be seen that LinearSVC and Logistic Regression working better than MultinomialNB and RandomForestClassifier
# let's try boosting and ensembling with top 3 performers

X_train, X_test, y_train, y_test, indices_train, indices_test = train_test_split(features, labels, 
                                                                                 df.index, test_size=0.25, random_state=0)


# let's have our own ensembles method
# a combination 

# votoing hard means votes on the labels not on the probabilities
ensemble_vc = VotingClassifier( estimators=[ ('adaboost', adaboost) , 
                                            ('random_forest', random_forest) , 
                                            ('bg', bg)] , voting='hard' )

svm_poly_2 = SVC(kernel='poly' , degree= 2)
svm_poly_2.fit(X_train, y_train)

# trying different kinds of SVM, 
svm_rbf = SVC(kernel='rbf')
svm_rbf.fit(X_train, y_train)

# trying different kinds of SVM, 
svm_sigmoid = SVC(kernel='sigmoid')
svm_sigmoid.fit(X_train, y_train)

# chosing Linear SVC and working on it
linear_svm = LinearSVC()
linear_svm.fit(X_train, y_train)
#fig, ax = plt.subplots(figsize=(10,10))
#sns.heatmap(conf_mat, annot=True, fmt='d',
#            xticklabels=collection_labels_df.collection.values, yticklabels=collection_labels_df.collection.values)
#plt.ylabel('Actual')
#plt.xlabel('Predicted')
#plt.show()

y_pred = linear_svm.predict(X_test)

#conf_mat = confusion_matrix(y_test, y_pred)


print(metrics.classification_report(y_test, y_pred, 
                                    target_names=df['collection'].unique()))
