==> 5a-numeric-text-features-combined-modeling.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns

# ml models decistion tree
from sklearn.tree import DecisionTreeClassifier

# One of the Ensemle model -- Random Forset which is ensemple of decision trees
from sklearn.ensemble import RandomForestClassifier, BaggingClassifier, AdaBoostClassifier, VotingClassifier

# import logistic regression, SVC  for ensembling
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
# Naive Bayes
from sklearn.naive_bayes import MultinomialNB

# split the data into test and train
from sklearn.model_selection import train_test_split

# packages for metric for evalaution of the models
from sklearn.metrics import classification_report , accuracy_score
from sklearn import metrics
# to save the models so that we don't need to train the models every time we need to do prediction
from sklearn.externals import joblib



df_numeric = pd.read_csv(r'datasets\1b-crest-after-numeric-preprocessing.csv')
df_text = pd.read_csv(r'datasets\1a-crest-after-text-cleaning.csv')

df_numeric.columns

df_numeric.head(5)

df_text.columns

# stack the DataFrames on top of each other/ Merge datasets row -wise 
df1 = df_numeric[['document_page_count', 'sequence_number', 'collection_labels', 'publication_day', 'publication_month', 'publication_year',
       'check_digit', 'release_decision_num', 'content_type_num', 'sub_department_num']]

df2 = df_text[['title_word_count', 'avg_word_len']]

# concat column wise so axis would 1
df = pd.concat([df1, df2], axis=1)
df.shape

df.to_csv(r'datasets\5a_CREST_text_and_numeric_features.csv' , index=False)

df.columns

    # use resample method from scikit-learn
    from sklearn.utils import resample
   
df_for_production = resample(df, replace=True,    # sample with replacement
                                 n_samples=5,     # to match number of values in each class
                                 random_state=123) # reproducible results

df_for_production.to_csv(r'datasets\CREST_5_samples_for_production.csv' , index=False)

X = df[['document_page_count', 'sequence_number', 'publication_day', 'publication_month', 'publication_year',
       'check_digit', 'release_decision_num', 'content_type_num',
       'sub_department_num', 'title_word_count', 'avg_word_len']]

y = df[['collection_labels']]

X.head(2)

y.head(2)

X_train, X_test, y_train, y_test = train_test_split(X , y, test_size = 0.2 , random_state = 4)

# let's try how ensemling ( Random forest wit 10 decision trees ) performs 
random_forest = RandomForestClassifier(n_estimators=10)
random_forest.fit(X_train, y_train)

# save the  random model
filepath = r'saved-models-joblib\trained-text-and-numeric-features\random_forest_model'
joblib.dump(random_forest ,filepath )
# load model for prediction
random_forest = joblib.load(filepath)

print(" Random Forest Accuracy = ")
print(random_forest.score(X_test, y_test))
print("\n")
y_pred_rf = random_forest.predict(X_test)
print(classification_report(y_test, y_pred_rf, target_names=df_numeric['collection'].unique()))

# Random Forest is performing better than Decision tree.Ensemling improved precision and F1-score.
# let's try bagging classifier
# We will give 20 Decisions Trees, each with 50% of training dataset and 100% features
bg = BaggingClassifier( DecisionTreeClassifier() , max_samples = 0.5 , max_features= 1.0 , n_estimators=20 )
bg.fit(X_train , y_train)

# save the baggign classifier model
filepath = r'saved-models-joblib\trained-text-and-numeric-features\BaggingClassifier_model'
joblib.dump(bg ,filepath )
# load model for prediction
bg = joblib.load(filepath)

print(" Bagging Classifier Accuracy = ")
print(bg.score(X_test, y_test))
print("\n")

y_pred = bg.predict(X_test)
print(classification_report(y_test, y_pred, target_names=df_numeric['collection'].unique()))

# let's see the performance of BoostingClassifier -ADABoost
# 10 decision trees , learning rate will shrink the contribution of each individual learner 
adaboost = AdaBoostClassifier( DecisionTreeClassifier() , n_estimators= 10 , learning_rate= 1 )
adaboost.fit(X_train, y_train)

# save the adaboost classifier model
filepath = r'saved-models-joblib\trained-text-and-numeric-features\adaboost_model'
joblib.dump(adaboost ,filepath )
# load model for prediction
adaboost = joblib.load(filepath)

print("ADA Boost model Accuracy = ")
print(adaboost.score(X_test , y_test))

print("\n")

y_pred_adaboost = adaboost.predict(X_test)
print(classification_report(y_test , y_pred_adaboost, target_names=df_numeric['collection'].unique()))

# let's have our own ensembles method

# votoing hard means votes on the labels not on the probabilities
ensemble_vc = VotingClassifier( estimators=[ ('adaboost', adaboost) , 
                                            ('random_forest', random_forest) , 
                                            ('bg', bg)] , voting='hard' )

ensemble_vc.fit(X_train , y_train)

# save the custom ensembling classifier model 
filepath = r'saved-models-joblib\trained-text-and-numeric-features\ensemble_voting_classifier_model'
joblib.dump(ensemble_vc ,filepath )
# load model for prediction
ensemble_vc = joblib.load(filepath)

print("Custom Voting Classifier Accuracy = ")
print(ensemble_vc.score(X_test, y_test))
print("\n")

# print the metrices for our ensemmbled model
y_pred_ensemble_vc = ensemble_vc.predict(X_test)
print( classification_report(y_test , y_pred_ensemble_vc, target_names=df_numeric['collection'].unique()))
